move the rest of the the tests to the sshConfig file
